ChangeSet with Pizza Price int

+++ src/main/java/com/example/sbertaste/model/PizzaEntity.java
@@ -2,7 +2,10 @@ package com.example.sbertaste.model;
 
 import lombok.*;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.SequenceGenerator;
+import javax.persistence.Table;
 
 @Entity
 @Table(name = "pizza")
@@ -19,6 +22,6 @@ public class PizzaEntity extends CommonEntity {
     private String name;
 
     @Column(name = "price")
-    private double price;
+    private int price;
 
 }
